Fix more circular dependencies and readd circular dependency plugin
