+++ frontend/app/config/webpack.config.js
@@ -41,7 +41,7 @@ const ForkTsCheckerWebpackPlugin =
 const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin")
 const createEnvironmentHash = require("./webpack/persistentCache/createEnvironmentHash")
 // reenable when circular dependencies are going to be fixed
-// const CircularDependencyPlugin = require("circular-dependency-plugin")
+const CircularDependencyPlugin = require("circular-dependency-plugin")
 
 // Source maps are resource heavy and can cause out of memory issue for large source files.
 const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== "false"
@@ -802,15 +802,15 @@ module.exports = function (webpackEnv) {
           },
         }),
       // reenable when circular dependencies are going to be fixed
-      // new CircularDependencyPlugin({
-      //   // exclude detection of files based on a RegExp
-      //   exclude: /node_modules/,
-      //   // add errors to webpack instead of warnings
-      //   failOnError: false,
-      //   // allow import cycles that include an asyncronous import,
-      //   // e.g. via import(/* webpackMode: "weak" */ './file.js')
-      //   allowAsyncCycles: false,
-      // }),
+      new CircularDependencyPlugin({
+        // exclude detection of files based on a RegExp
+        exclude: /node_modules/,
+        // add errors to webpack instead of warnings
+        failOnError: false,
+        // allow import cycles that include an asyncronous import,
+        // e.g. via import(/* webpackMode: "weak" */ './file.js')
+        allowAsyncCycles: false,
+      }),
     ].filter(Boolean),
     // Turn off performance processing because we utilize
     // our own hints via the FileSizeReporter
