+++ frontend/lib/src/components/shared/StreamlitMarkdown/StreamlitMarkdown.test.tsx
@@ -20,22 +20,14 @@ import ReactMarkdown from "react-markdown"
 import { mount, render } from "src/lib/test_util"
 import { cleanup } from "@testing-library/react"
 import IsSidebarContext from "src/components/core/Sidebar/IsSidebarContext"
-import { Heading as HeadingProto } from "src/autogen/proto"
 import { colors } from "src/theme/primitives/colors"
 import StreamlitMarkdown, {
   LinkWithTargetBlank,
   createAnchorFromText,
   HeadingWithAnchor,
-  Heading,
-  HeadingProtoProps,
   CustomCodeTag,
   CustomCodeTagProps,
 } from "./StreamlitMarkdown"
-
-import {
-  InlineTooltipIcon,
-  StyledLabelHelpWrapper,
-} from "src/components/shared/TooltipIcon"
 
 import { StyledLinkIconContainer } from "./styled-components"
 
@@ -287,98 +279,6 @@ describe("StreamlitMarkdown", () => {
       // Removes rendered StreamlitMarkdown component before next case run
       cleanup()
     })
-  })
-})
-
-const getHeadingProps = (
-  elementProps: Partial<HeadingProto> = {}
-): HeadingProtoProps => ({
-  width: 5,
-  element: HeadingProto.create({
-    anchor: "some-anchor",
-    tag: "h1",
-    body: `hello world
-          this is a new line`,
-    ...elementProps,
-  }),
-})
-
-describe("Heading", () => {
-  it("renders properly after a new line", () => {
-    const props = getHeadingProps()
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual("hello world")
-    expect(wrapper.find("RenderedMarkdown").at(1).text()).toEqual(
-      "this is a new line"
-    )
-  })
-
-  it("renders properly without a new line", () => {
-    const props = getHeadingProps({ body: "hello" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual("hello")
-    expect(wrapper.find("StyledStreamlitMarkdown")).toHaveLength(1)
-  })
-
-  it("renders anchor link", () => {
-    const props = getHeadingProps({ body: "hello" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("StyledLinkIcon")).toHaveLength(1)
-  })
-
-  it("does not renders anchor link when it is hidden", () => {
-    const props = getHeadingProps({ body: "hello", hideAnchor: true })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("StyledLinkIcon")).toHaveLength(0)
-  })
-
-  it("renders properly with help text", () => {
-    const props = getHeadingProps({ body: "hello", help: "help text" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual("hello")
-    expect(wrapper.find("StyledStreamlitMarkdown")).toHaveLength(1)
-    expect(wrapper.find(StyledLabelHelpWrapper).exists()).toBe(true)
-    const inlineTooltipIcon = wrapper.find(InlineTooltipIcon)
-    expect(inlineTooltipIcon.exists()).toBe(true)
-    expect(inlineTooltipIcon.props().content).toBe("help text")
-  })
-
-  it("does not render ol block", () => {
-    const props = getHeadingProps({ body: "1) hello" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual("1) hello")
-    expect(wrapper.find("ol")).toHaveLength(0)
-  })
-
-  it("does not render ul block", () => {
-    const props = getHeadingProps({ body: "* hello" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual("* hello")
-    expect(wrapper.find("ul")).toHaveLength(0)
-  })
-
-  it("does not render blockquote with >", () => {
-    const props = getHeadingProps({ body: ">hello" })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual(">hello")
-    expect(wrapper.find("blockquote")).toHaveLength(0)
-  })
-
-  it("does not render tables", () => {
-    const props = getHeadingProps({
-      body: `| Syntax | Description |
-        | ----------- | ----------- |
-        | Header      | Title       |
-        | Paragraph   | Text        |`,
-    })
-    const wrapper = mount(<Heading {...props} />)
-    expect(wrapper.find("h1").text()).toEqual(`| Syntax | Description |`)
-    expect(wrapper.find("RenderedMarkdown").at(1).text()).toEqual(
-      `| ----------- | ----------- |
-    | Header      | Title       |
-    | Paragraph   | Text        |`
-    )
-    expect(wrapper.find("table")).toHaveLength(0)
   })
 })
