+++ frontend/lib/src/components/shared/StreamlitMarkdown/StreamlitMarkdown.tsx
@@ -53,11 +53,6 @@ import {
 } from "src/theme/index"
 
 import {
-  InlineTooltipIcon,
-  StyledLabelHelpWrapper,
-} from "src/components/shared/TooltipIcon"
-
-import {
   StyledHeaderContainer,
   StyledHeaderContent,
   StyledLinkIcon,
@@ -70,7 +65,7 @@ import { StyledCopyButtonContainer } from "src/components/elements/CodeBlock/sty
 import CopyButton from "src/components/elements/CodeBlock/CopyButton"
 import StreamlitSyntaxHighlighter from "src/components/elements/CodeBlock/StreamlitSyntaxHighlighter"
 
-enum Tags {
+export enum Tags {
   H1 = "h1",
   H2 = "h2",
   H3 = "h3",
@@ -465,86 +460,6 @@ export function LinkWithTargetBlank(props: LinkProps): ReactElement {
     >
       {children}
     </a>
-  )
-}
-
-function makeMarkdownHeading(tag: string, markdown: string): string {
-  switch (tag.toLowerCase()) {
-    case Tags.H1: {
-      return `# ${markdown}`
-    }
-    case Tags.H2: {
-      return `## ${markdown}`
-    }
-    case Tags.H3: {
-      return `### ${markdown}`
-    }
-    default: {
-      throw new Error(`Unrecognized tag for header: ${tag}`)
-    }
-  }
-}
-
-export function Heading(props: HeadingProtoProps): ReactElement {
-  const { width, element } = props
-  const { tag, anchor, body, help, hideAnchor } = element
-  const isSidebar = React.useContext(IsSidebarContext)
-  // st.header can contain new lines which are just interpreted as new
-  // markdown to be rendered as such.
-  const [heading, ...rest] = body.split("\n")
-
-  return (
-    <div className="stMarkdown" style={{ width }}>
-      <StyledStreamlitMarkdown
-        isCaption={Boolean(false)}
-        isInSidebar={isSidebar}
-        style={{ width }}
-        data-testid="stMarkdownContainer"
-      >
-        <StyledHeaderContainer>
-          <HeadingWithAnchor tag={tag} anchor={anchor} hideAnchor={hideAnchor}>
-            {help ? (
-              <StyledLabelHelpWrapper>
-                <RenderedMarkdown
-                  source={makeMarkdownHeading(tag, heading)}
-                  allowHTML={false}
-                  // this is purely an inline string
-                  overrideComponents={{
-                    p: Fragment,
-                    h1: Fragment,
-                    h2: Fragment,
-                    h3: Fragment,
-                    h4: Fragment,
-                    h5: Fragment,
-                    h6: Fragment,
-                  }}
-                />
-                <InlineTooltipIcon content={help} />
-              </StyledLabelHelpWrapper>
-            ) : (
-              <RenderedMarkdown
-                source={makeMarkdownHeading(tag, heading)}
-                allowHTML={false}
-                // this is purely an inline string
-                overrideComponents={{
-                  p: Fragment,
-                  h1: Fragment,
-                  h2: Fragment,
-                  h3: Fragment,
-                  h4: Fragment,
-                  h5: Fragment,
-                  h6: Fragment,
-                }}
-              />
-            )}
-          </HeadingWithAnchor>
-        </StyledHeaderContainer>
-        {/* Only the first line of the body is used as a heading, the remaining text is added as regular mardkown below. */}
-        {rest.length > 0 && (
-          <RenderedMarkdown source={rest.join("\n")} allowHTML={false} />
-        )}
-      </StyledStreamlitMarkdown>
-    </div>
   )
 }
